http: handle remote connection close
We don't really understand Connection: close and thus don't find out
when the remote decided not to reuse the connection. Let's just try once
more if we encounter an EOF in a HTTP connection.

Also, always use a separate connection for the tunnel. The tunnel
connection is long-running and not really restartable so we can't really
just re-do it on EOF. Besides that, always close it when the tunnel
closes and don't use it for the /remote/logout request -- the server can't
tell that we're no longer doing PPP and are back to HTTP.